requirements.txt
python-dotenv
azure-identity
azure-storage-blob
azure-core
azure-search-documents
openai
tiktoken